feat(webauthn): permit Backup Eligibility flag validation to be relaxed
From commit 2195f33 onwards, the value of the Backup Eligibility (BE)
flag in a credential is required not to change between registration and
use. As noted in the commit message, this breaks compatibility with
client implementations that do not strictly adhere to the WebAuthn Level
3 specification.

Add a configuration option `RelaxBEFlagValidation` that allows this
check to be relaxed.